requirements.txt
Django>=4.2
celery
django-celery-beat
redis
psycopg2-binary
beautifulsoup4
requests
python-dotenv>=0.21.0
djangorestframework>=3.12
Pillow
djangorestframework-simplejwt
coverage
